Remove circular dependency between user service and authentication
