+++ src/main/java/ru/solution/test_task_for_gitflic_team/service/UserService.java
@@ -5,5 +5,4 @@ import ru.solution.test_task_for_gitflic_team.entity.User;
 
 public interface UserService extends UserDetailsService {
     User register(String username, String password);
-    User login(String username, String password);
 }
